extract MothIcon to shared ui component
